stock/requirements.txt
sqlalchemy
tushare 
pymysql 
requests 
lxml 
bs4 
numpy 
pandas 
scipy
matplotlib 
openpyxl
xlwt
xlrd 
jupyterlab
mysql-connector-python
